requirements.txt
google-generativeai
elevenlabs
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-classic
gradio
requests
